Fixes #31740 - Refactor Katello::EventDaemon   to fix intermittent ping failures
  -move EventDaemon to new runner namespace
  -move Rails cache reads to EventDaemon::Runner
  -move Rails cache writes to EventDaemon::Monitor
  -move the two files above from app/services to app/lib
  -Don't clear cache on service stop
  -update tests
  -add test for cache write
